backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ucabaur <ambroisel@theodo.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0"
jupyter = "^1.0.0"
pandas = "^2.1.4"
numpy = "^1.26.2"
dj-database-url = "^2.1.0"
psycopg = "^3.1.16"
python-dotenv = "^1.0.0"
django-extensions = "^3.2.3"
werkzeug = "^3.0.1"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
django-import-export = "^3.3.5"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
isort = "^5.13.2"
black = "^23.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
